// src/controllers/relationships/types/RelationType.ts
export enum RelationType {
    CompanyHasBrand = 'has-brand',
    CompanyHasImage = 'has-image',
    CompanyHasPrimeImage = 'has-prime-image',
    BrandBelongsToCompany = 'belongs-to-company',
    BrandHasCarModel = 'has-car-model',
    BrandHasImage = 'has-image',
    BrandHasPrimeImage = 'has-prime-image',
    CarModelBelongsToBrand = 'belongs-to-brand',
    CarModelHasSuccessor = 'has-successor',
    CarModelIsSuccessorOf = 'is-successor-of',
    CarModelHasVariant = 'has-variant',
    CarModelHasImage = 'has-image',
    CarModelHasPrimeImage = 'has-prime-image',
    CarModelVariantIsVariantOf = 'is-variant-of',
    CarModelVariantAchievedSessionResult = 'achieved-session-result',
    CarModelVariantAchievedLapTime = 'achieved-lap-time',
    CarModelVariantIsFeaturedInRacingGame = 'is-featured-in-racing-game',
    CarModelVariantHasImage = 'has-image',
    CarModelVariantHasPrimeImage = 'has-prime-image',
    RaceTrackHasLayout = 'has-layout',
    RaceTrackHostedRacingEvent = 'hosted-racing-event',
    RaceTrackHasImage = 'has-image',
    RaceTrackHasPrimeImage = 'has-prime-image',
    TrackLayoutBelongsToRaceTrack = 'belongs-to-race-track',
    TrackLayoutWasUsedByRacingEvent = 'was-used-by-racing-event',
    TrackLayoutHasLapTime = 'has-lap-time',
    TrackLayoutIsFeaturedInRacingGame = 'is-featured-in-racing-game',
    TrackLayoutHasImage = 'has-image',
    TrackLayoutHasPrimeImage = 'has-prime-image',
    RacingSeriesHasRacingEvent = 'has-racing-event',
    RacingSeriesHasImage = 'has-image',
    RacingSeriesHasPrimeImage = 'has-prime-image',
    RacingEventBelongsToRacingSeries = 'belongs-to-racing-series',
    RacingEventIsFollowedByEvent = 'is-followed-by-event',
    RacingEventFollowsEvent = 'follows-event',
    RacingEventTookPlaceAtRaceTrack = 'took-place-at-race-track',
    RacingEventUsedTheTrackLayout = 'used-the-track-layout',
    RacingEventHasRacingSession = 'has-racing-session',
    RacingEventHasImage = 'has-image',
    RacingEventHasPrimeImage = 'has-prime-image',
    RacingSessionBelongsToRacingEvent = 'belongs-to-racing-event',
    RacingSessionHasSessionResult = 'has-session-result',
    RacingSessionHasImage = 'has-image',
    RacingSessionHasPrimeImage = 'has-prime-image',
    SessionResultBelongsToRacingSession = 'belongs-to-racing-session',
    SessionResultHasLapTime = 'has-lap-time',
    SessionResultAchievedWithCarModelVariant = 'achieved-with-car-model-variant',
    SessionResultHasImage = 'has-image',
    SessionResultHasPrimeImage = 'has-prime-image',
    LapTimeBelongsToSessionResult = 'belongs-to-session-result',
    LapTimeAchievedOnTrackLayout = 'achieved-on-track-layout',
    LapTimeAchievedWithCarModelVariant = 'achieved-with-car-model-variant',
    LapTimeHasImage = 'has-image',
    LapTimeHasPrimeImage = 'has-prime-image',
    RacingGameFeaturesCarModelVariant = 'features-car-model-variant',
    RacingGameFeaturesTrackLayout = 'features-track-layout',
    RacingGameReleasedOnGamingPlatform = 'released-on-gaming-platform',
    RacingGameHasImage = 'has-image',
    RacingGameHasPrimeImage = 'has-prime-image',
    GamingPlatformFeaturesRacingGame = 'features-racing-game',
    GamingPlatformHasImage = 'has-image',
    GamingPlatformHasPrimeImage = 'has-prime-image',
    ImageBelongsToNode = 'belongs-to-node',
    ImageIsPrimeImageOfNode = 'is-prime-image-of-node',
}